Added Game of Catz Menu Project
The purpose of the gameofcatz menu module is to load input file and let users to find, add, remove or update their nodes and edges and moves through a world made of those nodes and edges. The menu contains 10 options for the user to select such as to allow the user to load the data that they want as well as to display and save the data in to file. The simulation mode has not been implemented so will be omitted from this description.